Separate folder for experience drawables
