+++ UIInfoSuite2/UIElements/ExperienceBar.cs
@@ -13,6 +13,7 @@ using System.Threading.Tasks;
 using UIInfoSuite2.Compatibility;
 using UIInfoSuite2.Infrastructure;
 using UIInfoSuite2.Infrastructure.Extensions;
+using UIInfoSuite2.UIElements.ExperienceElements;
 
 namespace UIInfoSuite2.UIElements
 {
